chore #50: 파일 분리
