+++ Projects/Features/Sources/MorePersonality/MorePersonalityView.swift
@@ -8,52 +8,8 @@
 
 import SwiftUI
 import ComposableArchitecture
+import Domain
 import DSKit
-
-struct Personality: Equatable, Identifiable {
-    let id = UUID()
-    let name: String
-    let keyword: String
-    let averageScore: Float
-}
-
-struct MorePersonalityFeature: Reducer {
-    public struct State: Equatable {
-        var personalities: [Personality]
-        
-        public init() {
-            self.personalities = [
-                Personality(name: "닉네임", keyword: "꼰대력", averageScore: 3.5)
-            ]
-        }
-    }
-    public enum Action: Equatable {
-        case loadPersonality
-        case savePersonality([Personality])
-    }
-    
-    public var body: some ReducerOf<Self> {
-        Reduce { state, action in
-            switch action {
-            case .loadPersonality:
-                return .run { send in
-                    try await Task.sleep(until: .now + .seconds(0.5), clock: .continuous)
-                    await send(.savePersonality([
-                        Personality(name: "닉네임", keyword: "꼰대력", averageScore: 3.5),
-                        Personality(name: "닉네임", keyword: "꼰대력", averageScore: 3.5),
-                        Personality(name: "닉네임", keyword: "꼰대력", averageScore: 3.5),
-                        Personality(name: "닉네임", keyword: "꼰대력", averageScore: 3.5),
-                        Personality(name: "닉네임", keyword: "꼰대력", averageScore: 3.5)
-                    ]))
-                }
-                
-            case .savePersonality(let data):
-                state.personalities.append(contentsOf: data)
-            }
-            return .none
-        }
-    }
-}
 
 struct MorePersonalityView: View {
     let store: StoreOf<MorePersonalityFeature>
